requirements.txt
cachetools==5.5.2
certifi
cffi==1.16.0
charset-normalizer
click
colorama
cryptography
et-xmlfile
Flask==2.3.3
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
googleapis-common-protos
httplib2
idna
itsdangerous
jinja2
MarkupSafe
numpy==2.1.3
oauth2client
oauthlib
openpyxl
pandas==2.2.3
proto-plus
protobuf
pyasn1
pyasn1-modules
pycparser
PyDrive2
pyopenssl
pyparsing
python-dateutil
pytz
PyYAML
requests
requests-oauthlib
rsa
six
typing-extensions
uritemplate
urllib3
Werkzeug==3.0.3
zipp
gunicorn

